chore: make separate VerboseHelperText component
